LPS-94286 Move test to correct file
