+++ modules/apps/asset/asset-list-test/src/testIntegration/java/com/liferay/asset/list/service/test/AssetListEntryAssetEntryRelServiceTest.java
@@ -16,6 +16,7 @@ package com.liferay.asset.list.service.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.asset.kernel.model.AssetEntry;
+import com.liferay.asset.list.exception.AssetListEntryAssetEntryRelPostionException;
 import com.liferay.asset.list.model.AssetListEntry;
 import com.liferay.asset.list.model.AssetListEntryAssetEntryRel;
 import com.liferay.asset.list.service.AssetListEntryAssetEntryRelLocalServiceUtil;
@@ -29,6 +30,7 @@ import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.kernel.test.util.GroupTestUtil;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
+import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
 import com.liferay.portal.kernel.transaction.Propagation;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
@@ -127,6 +129,32 @@ public class AssetListEntryAssetEntryRelServiceTest {
 			assetListEntryRelLocal.getAssetListEntryId());
 		Assert.assertEquals(
 			assetEntry.getEntryId(), assetListEntryRelLocal.getAssetEntryId());
+	}
+
+	@Test(expected = AssetListEntryAssetEntryRelPostionException.class)
+	public void testAddAssetListEntryAssetEntryRelWrongPosition()
+		throws Exception {
+
+		AssetListEntry assetListEntry = _addAssetListEntry("Asset List Title");
+
+		AssetEntry assetEntry = AssetTestUtil.addAssetEntry(
+			_group.getGroupId());
+
+		ServiceContext serviceContext =
+			ServiceContextTestUtil.getServiceContext(
+				_group.getGroupId(), TestPropsValues.getUserId());
+
+		long segmentsEntryId = RandomTestUtil.nextLong();
+
+		AssetListEntryAssetEntryRelLocalServiceUtil.
+			addAssetListEntryAssetEntryRel(
+				assetListEntry.getAssetListEntryId(), segmentsEntryId,
+				assetEntry.getEntryId(), 1, serviceContext);
+
+		AssetListEntryAssetEntryRelLocalServiceUtil.
+			addAssetListEntryAssetEntryRel(
+				assetListEntry.getAssetListEntryId(), segmentsEntryId,
+				assetEntry.getEntryId(), serviceContext);
 	}
 
 	@Test
